+++ gnc/pmc_sim/package.xml
@@ -1,0 +1,24 @@
+<?xml version="1.0"?>
+<package>
+  <name>pmc_sim</name>
+  <version>0.0.0</version>
+  <description>
+    Simulates the propulsion modules.
+  </description>
+  <license>
+  	Apache License, Version 2.0
+  </license>
+  <author email="astrobee-fsw@nx.arc.nasa.gov">
+    Astrobee Flight Software
+  </author>
+  <maintainer email="astrobee-fsw@nx.arc.nasa.gov">
+    Astrobee Flight Software
+  </maintainer>
+  <buildtool_depend>catkin</buildtool_depend>
+  <build_depend>roscpp</build_depend>
+  <build_depend>config_reader</build_depend>
+  <build_depend>gnc_autocode</build_depend>
+  <run_depend>roscpp</run_depend>
+  <run_depend>config_reader</run_depend>
+  <run_depend>gnc_autocode</run_depend>
+</package>
