+++ gnc/pmc_sim/readme.md
@@ -1,0 +1,5 @@
+\page pmc\_sim PMC Simulator
+
+The PMC simulator simulates the robot's propusion system. It is essentially the reverse of the
+FAM--- given impeller commands and servo commands, it returns the force and torque applied to the robot.
+
